rename file following pep8 rules
