+++ EduNLP/ModelZoo/DisenQNet/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 
-from .DisenQNet import DisenQNet
+from .disenqnet import DisenQNet
