--- EduNLP/Vector/DisenQModel/__init__.py
@@ -1,1 +1,0 @@
-from .DisenQModel import DisenQModel
