+++ EduNLP/Vector/disenqnet/__init__.py
@@ -1,0 +1,1 @@
+from .disenqnet import DisenQModel
